Preserve initctl while running chroot_(u)mount_install.sh directly

+++ chroot_umount_install.sh
@@ -2,8 +2,10 @@ export HOME=/root
 export LC_ALL=C
 
 chroot chroot rm -f /var/lib/dbus/machine-id
-chroot chroot rm -f /sbin/initctl
-chroot chroot dpkg-divert --rename --remove /sbin/initctl
+if [ -f "/chroot/tmp/initctl_divert" ]; then
+  chroot chroot rm -f /sbin/initctl
+  chroot chroot dpkg-divert --rename --remove /sbin/initctl
+fi
 chroot chroot apt-get clean
 chroot chroot rm -rf /tmp/*
 chroot chroot rm /etc/resolv.conf
